backupfirst run/bot/requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
python-dotenv==1.0.1
structlog==24.1.0